// src/app/Models/PackageModel/package-model.ts
import { ItineraryModel } from './itinerary-model';

export class PackageModel {
    PackageID: number;
    PackageType: string;
    PackageName: string;
    Days: number;
    ProfitPercentage: number;
    Price: number;
    Summary: string;
    NumberAvailable: number;
    MinPeople: number;
    MaxPeople: number;
    Image: string;
    Rating: number;
}